requirements.txt
flask==3.1.0
anthropic==0.52.0
openai==1.82.0
httpx==0.28.1
python-dotenv==1.1.0
